refactor(geo): merge array extension files
